# asm/amd64: optimizes initializeNodesForEncoding